Refactored & fixed tests for Conversion, *Context
Also got rid of the near-unused CommandStrings.
